Refactor homepage into utils and components
