+++ utils/index.js
@@ -1,0 +1,30 @@
+import * as prismicH from "@prismicio/helpers";
+
+export const dateFormatter = new Intl.DateTimeFormat("en-US", {
+  month: "short",
+  day: "numeric",
+  year: "numeric",
+});
+
+export const findFirstImage = (slices) => {
+  const imageSlice = slices.find((slice) => slice.slice_type === "image");
+
+  if (imageSlice && prismicH.isFilled.image(imageSlice.primary.image)) {
+    return imageSlice.primary.image;
+  }
+};
+
+export const getExcerpt = (slices) => {
+  const text = slices
+    .filter((slice) => slice.slice_type === "text")
+    .map((slice) => prismicH.asText(slice.primary.text))
+    .join(" ");
+
+  const excerpt = text.substring(0, 300);
+
+  if (text.length > 300) {
+    return excerpt.substring(0, excerpt.lastIndexOf(" ")) + "…";
+  } else {
+    return excerpt;
+  }
+};
